@@ -2,6 +2,7 @@
     "name": "sparkpost/php-sparkpost",
     "description": "SDK for interfacing with SparkPost APIs",
     "license": "Apache 2.0",
+    "version": "0.2.0",
     "authors": [
         {
             "name": "Message Systems, Inc."
@@ -9,11 +10,12 @@
     ],
     "minimum-stability": "stable",
     "require": {
-    	"php": ">=5.3.0",
-        "guzzlehttp/guzzle": "3.8.1"
+    	"php": ">=5.5.0",
+      "egeloen/http-adapter": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.3.*"
+        "phpunit/phpunit": "4.3.*",
+        "guzzlehttp/guzzle": "6.*"
     },
     "autoload": {
     	"psr-4": {
